Move PackageEventListner to BufferIOListener
