+++ src/Provider/ExtensionServiceProvider.php
@@ -3,7 +3,7 @@
 namespace Bolt\Provider;
 
 use Bolt\Composer\Action;
-use Bolt\Composer\PackageEventListener;
+use Bolt\Composer\EventListener\BufferIOListener;
 use Bolt\Composer\PackageManager;
 use Bolt\Extensions;
 use Bolt\Extensions\ExtensionsInfoService;
@@ -56,7 +56,7 @@ class ExtensionServiceProvider implements ServiceProviderInterface
 
         $app['extend.listener'] = $app->share(
             function ($app) {
-                return new PackageEventListener($app['extend.manager']);
+                return new BufferIOListener($app['extend.manager']);
             }
         );
